+++ src/test/java/woowacourse/common/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.acceptance;
+package woowacourse.common;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
